pyproject.toml
[project]
name = "fanllm"
version = "0.1.0"
description = "Fire one prompt at multiple LLMs in parallel, compare responses side by side."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Yazar", email = "yazararme@gmail.com" }]
keywords = ["llm", "openai", "anthropic", "google", "perplexity", "deepseek", "grok", "xai", "mistral", "comparison"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["rich>=13.0", "pytest>=8.0", "pytest-asyncio>=0.23"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
